Tap Card - Award Status Handling
Adds handling for award status of a beer and its representation on the
overview page. Beers with and award will have a medal on their tap card.
Adjusts fillInBeerData method to include global replacement so that any
instance of a key replacement phrase will be replaced with the
appropriate text. Consolidates the pint-image svg files into a single
HTML template that includes the new award icon HTML. Adds icons.css link
to the overview page. Refactors variable names that reference SVG's to
reference pintImageTemplate instead.